requirements.txt
Flask==2.3.5
Pillow==10.1.0
Werkzeug==2.3.7